configuration part modified!
In this version of code i modified the configuration part and solved the problems which occurred in the previous code version .Since the program was initially based on the time sequence without configuration period, the old  version would show flashing screen due to a random state machine when configuring the sensor. I changed the code and solved this problem.